The Enrichment is more of a response requirement than a business requirement. Hence moving it into DTO
